+++ app/Http/Controllers/RuanganController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Ruangan;
 use App\Models\JenisRuangan;
+use Illuminate\Support\Facades\Gate;
 
 class RuanganController extends Controller
 {
@@ -31,12 +32,16 @@ class RuanganController extends Controller
      */
     public function create()
     {
-        $data = [
-            'title' => 'Tambah Ruangan',
-            'jenis_ruangan' => JenisRuangan::all()
-        ];
+        if (Gate::allows('isAdmin')) {
+            $data = [
+                'title' => 'Tambah Ruangan',
+                'jenis_ruangan' => JenisRuangan::all()
+            ];
 
-        return view('dashboard.ruangan.create', $data);
+            return view('dashboard.ruangan.create', $data);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -47,19 +52,23 @@ class RuanganController extends Controller
      */
     public function store(Request $request)
     {
-        $messages = [
-            'no_ruangan.required' => 'No Ruangan harus diisi',
-            'jenis_ruangan_id.required' => 'Jenis Ruangan harus diisi'
-        ];
+        if (Gate::allows('isAdmin')) {
+            $messages = [
+                'no_ruangan.required' => 'No Ruangan harus diisi',
+                'jenis_ruangan_id.required' => 'Jenis Ruangan harus diisi'
+            ];
 
-        $request->validate([
-            'no_ruangan' => 'required',
-            'jenis_ruangan_id' => 'required'
-        ], $messages);
+            $request->validate([
+                'no_ruangan' => 'required',
+                'jenis_ruangan_id' => 'required'
+            ], $messages);
 
-        Ruangan::create($request->all());
+            Ruangan::create($request->all());
 
-        return redirect()->route('ruangan.index')->with('success', 'Ruangan berhasil ditambahkan');
+            return redirect()->route('ruangan.index')->with('success', 'Ruangan berhasil ditambahkan');
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -81,13 +90,17 @@ class RuanganController extends Controller
      */
     public function edit($id)
     {
-        $data = [
-            'title' => 'Edit Ruangan',
-            'ruangan' => Ruangan::findOrFail($id),
-            'jenis_ruangan' => JenisRuangan::all()
-        ];
+        if (Gate::allows('isAdmin')) {
+            $data = [
+                'title' => 'Edit Ruangan',
+                'ruangan' => Ruangan::findOrFail($id),
+                'jenis_ruangan' => JenisRuangan::all()
+            ];
 
-        return view('dashboard.ruangan.edit', $data);
+            return view('dashboard.ruangan.edit', $data);
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -99,19 +112,23 @@ class RuanganController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $messages = [
-            'no_ruangan.required' => 'No Ruangan harus diisi',
-            'jenis_ruangan_id.required' => 'Jenis Ruangan harus diisi'
-        ];
+        if (Gate::allows('isAdmin')) {
+            $messages = [
+                'no_ruangan.required' => 'No Ruangan harus diisi',
+                'jenis_ruangan_id.required' => 'Jenis Ruangan harus diisi'
+            ];
 
-        $request->validate([
-            'no_ruangan' => 'required',
-            'jenis_ruangan_id' => 'required'
-        ], $messages);
+            $request->validate([
+                'no_ruangan' => 'required',
+                'jenis_ruangan_id' => 'required'
+            ], $messages);
 
-        Ruangan::findOrFail($id)->update($request->all());
+            Ruangan::findOrFail($id)->update($request->all());
 
-        return redirect()->route('ruangan.index')->with('success', 'Ruangan berhasil diubah');
+            return redirect()->route('ruangan.index')->with('success', 'Ruangan berhasil diubah');
+        } else {
+            abort(403);
+        }
     }
 
     /**
@@ -122,8 +139,12 @@ class RuanganController extends Controller
      */
     public function destroy($id)
     {
-        Ruangan::destroy($id);
+        if (Gate::allows('isAdmin')) {
+            Ruangan::destroy($id);
 
-        return redirect()->route('ruangan.index')->with('success', 'Ruangan berhasil dihapus');
+            return redirect()->route('ruangan.index')->with('success', 'Ruangan berhasil dihapus');
+        } else {
+            abort(403);
+        }
     }
 }
